turn ui into popup operator
